+++ libs/components-web-kkshow/src/lib/main/KkshowMainBestBroadcaster.tsx
@@ -9,8 +9,9 @@ import {
   Stack,
 } from '@chakra-ui/react';
 import MotionBox from '@project-lc/components-core/MotionBox';
-import { useKkshowMain } from '@project-lc/hooks';
+import { useKkshowMain, useLiveShoppingNowOnLive } from '@project-lc/hooks';
 import NextLink from 'next/link';
+import { useMemo } from 'react';
 import { Pagination, Scrollbar } from 'swiper';
 import { Swiper, SwiperSlide } from 'swiper/react';
 import KkshowMainTitle from './KkshowMainTitle';
@@ -41,6 +42,7 @@ export function KkshowMainBestBroadcaster(): JSX.Element | null {
                 style={{ maxWidth: 190, width: '45%', paddingBottom: 24 }}
               >
                 <BestBroadcasterItem
+                  broadcasterId={x.broadcasterId || undefined}
                   avatarUrl={x.profileImageUrl}
                   broadcasterName={x.nickname}
                   href={x.promotionPageLinkUrl}
@@ -55,11 +57,18 @@ export function KkshowMainBestBroadcaster(): JSX.Element | null {
 }
 
 interface BestBroadcasterItemProps {
+  broadcasterId?: number;
   broadcasterName: string;
   avatarUrl: string;
   href: string;
 }
 export function BestBroadcasterItem(props: BestBroadcasterItemProps): JSX.Element {
+  const { data: nowLiveList } = useLiveShoppingNowOnLive({
+    broadcasterId: props.broadcasterId,
+  });
+  const isNowLive = useMemo(() => {
+    return nowLiveList && nowLiveList.length > 0;
+  }, [nowLiveList]);
   return (
     <LinkBox outline="none">
       <Stack
@@ -72,19 +81,40 @@ export function BestBroadcasterItem(props: BestBroadcasterItemProps): JSX.Elemen
         role="group"
       >
         <Avatar
-          w={{ base: 120, md: 160 }}
-          h={{ base: 120, md: 160 }}
+          size="2xl"
+          width="100%"
+          height="100%"
+          maxW={{ base: 120, md: 160 }}
+          maxH={{ base: 120, md: 160 }}
           src={props.avatarUrl}
           cursor="pointer"
+          border={isNowLive ? '2px solid red' : undefined}
           _groupHover={{
-            transition: '0.1s',
-            outline: '4px solid',
+            transition: '0.15s',
+            outline: '2px solid',
             outlineColor: 'blue.400',
-            outlineOffset: 4,
+            outlineOffset: 2,
             boxShadow: 'xl',
           }}
+          pos="relative"
+          _after={
+            isNowLive
+              ? {
+                  color: 'white',
+                  backgroundColor: 'red',
+                  content: '"LIVE"',
+                  position: 'absolute',
+                  fontSize: 'xs',
+                  height: 4,
+                  bottom: 2,
+                  px: 2,
+                  rounded: 'sm',
+                  display: 'flex',
+                  justify: 'center',
+                }
+              : undefined
+          }
         />
-
         {props.href ? (
           <NextLink passHref href={props.href || '#'}>
             <LinkOverlay isExternal={props.href.includes('http')}>
