+++ libs/shared-types/src/lib/dto/kkshowBcList.dto.ts
@@ -1,10 +1,11 @@
 import { Type } from 'class-transformer';
-import { IsInt, IsNumber, IsString } from 'class-validator';
+import { IsInt, IsNumber, IsOptional, IsString } from 'class-validator';
 
 export class CreateKkshowBcListDto {
   @IsString() nickname: string;
   @IsString() profileImage: string;
   @IsString() href: string;
+  @Type(() => Number) @IsOptional() @IsInt() broadcasterId?: number;
 }
 
 export class DeleteKkshowBcListDto {
